レンダリング関係の型と関数を移動
